pyproject.toml
[tool.poetry]
name = "homeggs-api"
version = "0.1.0"
description = ""
authors = ["thomas <psychoid@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
pytest = "^8.3.3"
pillow = "^10.4.0"
djoser = "^2.2.3"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
